% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_and_forcing.R
\name{create_dates_forcing}
\alias{create_dates_forcing}
\title{Reading dates and focing from filename}
\usage{
create_dates_forcing(dir)
}
\arguments{
\item{dir}{dir with raw files}
}
\value{
a database with Date(POSIX) and forcing from filenames
}
\description{
Reading dates and focing from filename
}
