% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf_calculator.R
\name{ddf_calculator}
\alias{ddf_calculator}
\title{Historical DDF calculation}
\usage{
ddf_calculator(ddf_data)
}
\arguments{
\item{ddf_data}{A dataset with historical RCP}
}
\value{
Creates a data table with all variables
}
\description{
Historical DDF calculation
}
