% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_and_spatial.R
\name{readDT}
\alias{readDT}
\title{Data reading function}
\usage{
readDT(dir, na.rm = FALSE)
}
\arguments{
\item{dir}{dir() all raw files}

\item{na.rm}{FALSE by default}
}
\description{
Data reading function
}
