% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spatial_data.R
\name{readDT_Sp}
\alias{readDT_Sp}
\title{Spatial Data}
\usage{
readDT_Sp(dir, ext)
}
\arguments{
\item{dir}{dir with .nc files}

\item{ext}{coordinate system for the area}
}
\value{
A dataset with precip, lon, lat
}
\description{
Spatial Data
}
