% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_course.R
\name{spatial_function}
\alias{spatial_function}
\title{Spatial data lecture}
\usage{
spatial_function(dir, aux, dates_forcing)
}
\arguments{
\item{dir}{A directory of data}

\item{aux}{data with precipitation value and id}

\item{create_dates_forcing}{data with forcing and dates}
}
\value{
a dataset from spatial data course
}
\description{
Spatial data lecture
}
